+++ routes/MovieRoute.js
@@ -1,0 +1,6 @@
+const express = require("express");
+const router = express.Router();
+
+const MovieController = require("../controllers/MovieController");
+router.get("/genre",MovieController.getMovieByGenre);
+module.exports = router;
