--- routes/MoviesRoute.js
@@ -1,6 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-const MoviesController = require("../controllers/MoviesController");
-router.get("/genre",MoviesController.getMoviesByGenre);
-module.exports = router;
