+++ services/MovieService.js
@@ -1,0 +1,8 @@
+const MovieDao = require('../models/MovieDao')
+
+const getMovieByGenre= async (genreId) => {
+	return await MovieDao.getMovieByGenre(genreId)
+}
+
+
+module.exports = {getMovieByGenre}
